requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.4,<3.0
python-dotenv>=1.0,<2.0
weaviate-client>=4.20,<5.0
sentence-transformers>=3.0,<4.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0